requirements.txt
numpy
jupyter
ipykernel
nbterm
scipy
matplotlib
scikit-image
galois
sympy
tensorflow
pandas
dotenv
alembic
psycopg2
setuptools-scm
fastapi>=0.68.0,<0.69.0
fastapi-sqlalchemy
sqlalchemy-utils
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
gunicorn 
databases[postgresql]
python-jose[cryptography]
passlib[bcrypt]